+++ src/com/walmartlabs/system_viz.clj
@@ -4,7 +4,8 @@
     [com.stuartsierra.component :as component]
     [clojure.set :as set]
     [clojure.java.browse :refer [browse-url]]
-    [clojure.java.io :as io])
+    [clojure.java.io :as io]
+    [clojure.string :as str])
   (:import
     (java.io File)
     (java.util.concurrent TimeUnit)))
@@ -25,11 +26,11 @@
            (cond-let ~@more-forms))))))
 
 (defn ^:private system->dot
-  [system horizontal]
+  [system horizontal highlight-attrs]
   (println "digraph System {")
 
   (when horizontal
-    (println "graph [rankdir=LR]"))
+    (println "  graph [rankdir=LR]"))
 
   ;; Now find all the unknown keys (dependencies to unknown components).
   (let [all-keys (->> system
@@ -46,7 +47,22 @@
 
   (doseq [[component-key component] system
           :let [component-key' (quoted component-key)]]
-    (println (format "  %s;" component-key'))
+    (print " " component-key')
+    (when (map? component)
+      (let [color (:systemviz/color component)
+            terms (reduce (fn [terms [k v]]
+                            (conj terms (str (name k) "=" v)))
+                          []
+                          (cond-> (:systemviz/attrs component)
+                            color (assoc :color (name color)
+                                         :style 'filled)
+                            (:systemviz/highlight component) (merge highlight-attrs)))]
+        (when (seq terms)
+
+          (print (str " ["
+                      (str/join ", " terms)
+                      "]")))))
+    (println ";")
     (doseq [[local-key system-key] (component/dependencies component)]
       (print (format "  %s -> %s"
                      component-key'
@@ -66,6 +82,19 @@
   A temporary file is created to store the Graphviz
   program; this is then fed through the `dot`
   command line tool to generate a PDF file.
+
+  Visualization recognizes three special keys:
+
+  :graphviz/highlight
+  : Adds node attributes to highlight the component in
+    the graph. By default, skyblue color and 24 point font.
+
+  :graphviz/color
+  : Sets the color of the component to the provided color value,
+    and sets the style of the node to be filled.
+
+  :graphviz/attributes
+  : A map of keys and values; these are added as node attributes.
 
   The resulting PDF file may optionally be opened, or
   its location output to the console.
@@ -87,6 +116,10 @@
   :horizontal
   : If true (the default is false), then the image will be laid out
     horizontally instead of vertically.
+
+  :highlight
+  : A map of Graphviz attributes and values to be applied to components
+    where :graphviz/highlight is true.
     
   :save-as
   : If provided, then this is the path to which the graphviz source file 
@@ -98,13 +131,16 @@
    (visualize-system system nil))
   ([system options]
    (cond-let
-     :let [{:keys [format open save-as horizontal]
+     :let [{:keys [format open save-as horizontal highlight]
             :or {format :pdf
                  horizontal false
-                 open true}} options
+                 open true
+                 highlight {:color 'skyblue
+                            :style 'filled
+                            :fontsize 24}}} options
            format-name (name format)
            dot (with-out-str
-                 (system->dot system horizontal))
+                 (system->dot system horizontal highlight))
            gvfile (if (some? save-as)
                     (io/file save-as)
                     (File/createTempFile "system-" ".gv"))